+++ pystac/link.py
@@ -43,29 +43,35 @@ class Link:
         media_type : Optional description of the media type. Registered Media Types
             are preferred. See :class:`~pystac.MediaType` for common media types.
         title : Optional title for this link.
-        properties : Optional, additional properties for this link. This is used
+        extra_fields : Optional, additional fields for this link. This is used
             by extensions as a way to serialize and deserialize properties on link
             object JSON.
-
-    Attributes:
-        rel : The relation of the link (e.g. 'child', 'item'). Registered rel Types
-            are preferred. See :class:`~pystac.RelType` for common media types.
-        target : The target of the link. If the link is
-            unresolved, or the link is to something that is not a STACObject,
-            the target is an HREF. If resolved, the target is a STACObject.
-        media_type : Optional description of the media type.
-            Registered Media Types are preferred. See
-            :class:`~pystac.MediaType` for common media types.
-        title : Optional title for this link.
-        properties : Optional, additional properties for this link.
-            This is used by extensions as a way to serialize and deserialize properties
-            on link object JSON.
-        owner : The owner of this link. The link will use
-            its owner's root catalog
-            :class:`~pystac.resolved_object_cache.ResolvedObjectCache` to resolve
-            objects, and will create absolute HREFs from relative HREFs against
-            the owner's self HREF.
     """
+
+    rel: Union[str, pystac.RelType]
+    """The relation of the link (e.g. 'child', 'item'). Registered rel Types are
+    preferred. See :class:`~pystac.RelType` for common media types."""
+
+    target: Union[str, "STACObject_Type"]
+    """The target of the link. If the link is unresolved, or the link is to something
+    that is not a STACObject, the target is an HREF. If resolved, the target is a
+    STACObject."""
+
+    media_type: Optional[str]
+    """Optional description of the media type. Registered Media Types are preferred.
+    See :class:`~pystac.MediaType` for common media types."""
+
+    title: Optional[str]
+    """Optional title for this link."""
+
+    extra_fields: Dict[str, Any]
+    """Optional, additional fields for this link. This is used by extensions as a
+    way to serialize and deserialize properties on link object JSON."""
+
+    owner: Optional["STACObject_Type"]
+    """The owner of this link. The link will use its owner's root catalog
+    :class:`~pystac.resolved_object_cache.ResolvedObjectCache` to resolve objects, and
+    will create absolute HREFs from relative HREFs against the owner's self HREF."""
 
     def __init__(
         self,
@@ -73,14 +79,14 @@ class Link:
         target: Union[str, "STACObject_Type"],
         media_type: Optional[str] = None,
         title: Optional[str] = None,
-        properties: Optional[Dict[str, Any]] = None,
+        extra_fields: Optional[Dict[str, Any]] = None,
     ) -> None:
         self.rel = rel
-        self.target: Union[str, "STACObject_Type"] = target  # An object or an href
+        self.target = target
         self.media_type = media_type
         self.title = title
-        self.properties = properties
-        self.owner: Optional["STACObject_Type"] = None
+        self.extra_fields = extra_fields or {}
+        self.owner = None
 
     def set_owner(self, owner: Optional["STACObject_Type"]) -> "Link":
         """Sets the owner of this link.
@@ -254,9 +260,8 @@ class Link:
         if self.title is not None:
             d["title"] = self.title
 
-        if self.properties:
-            for k, v in self.properties.items():
-                d[k] = v
+        for k, v in self.extra_fields.items():
+            d[k] = v
 
         return d
 
@@ -293,16 +298,16 @@ class Link:
         media_type = d.pop("type", None)
         title = d.pop("title", None)
 
-        properties = None
+        extra_fields = None
         if any(d):
-            properties = d
+            extra_fields = d
 
         return cls(
             rel=rel,
             target=href,
             media_type=media_type,
             title=title,
-            properties=properties,
+            extra_fields=extra_fields,
         )
 
     @classmethod